Refactor SearchAttributeMenu
