+++ src/lib/components/workflow/filter-search/index.svelte
@@ -1,5 +1,5 @@
 <script lang="ts" context="module">
-  import { afterUpdate, noop } from 'svelte/internal';
+  import { afterUpdate } from 'svelte/internal';
   import { writable, type Writable } from 'svelte/store';
   import { fly } from 'svelte/transition';
 
@@ -21,19 +21,10 @@
 
   import WorkflowAdvancedSearch from '$lib/components/workflow/workflow-advanced-search.svelte';
   import Button from '$lib/holocene/button.svelte';
-  import Icon from '$lib/holocene/icon/icon.svelte';
-  import Input from '$lib/holocene/input/input.svelte';
-  import {
-    Menu,
-    MenuButton,
-    MenuContainer,
-    MenuItem,
-  } from '$lib/holocene/menu';
   import ToggleSwitch from '$lib/holocene/toggle-switch.svelte';
   import { translate } from '$lib/i18n/translate';
   import type { WorkflowFilter } from '$lib/models/workflow-filters';
   import { workflowFilters } from '$lib/stores/filters';
-  import { sortedSearchAttributeOptions } from '$lib/stores/search-attributes';
   import { refresh } from '$lib/stores/workflows';
   import {
     isBooleanFilter,
@@ -55,9 +46,9 @@
   import DateTimeFilter from './datetime-filter.svelte';
   import FilterList from './filter-list.svelte';
   import NumberFilter from './number-filter.svelte';
+  import SearchAttributeMenu from './search-attribute-menu.svelte';
   import StatusFilter from './status-filter.svelte';
   import TextFilter from './text-filter.svelte';
-
   const filter = writable<WorkflowFilter>(emptyFilter());
   const activeQueryIndex = writable<number>(null);
   const focusedElementId = writable<string>('');
@@ -100,7 +91,6 @@
       $workflowFilters = [...$workflowFilters, $filter];
     }
     filter.set(emptyFilter());
-    searchAttributeValue = '';
     onSearch();
   }
 
@@ -124,17 +114,6 @@
     return '';
   }
 
-  function isOptionDisabled(value: string, filters: WorkflowFilter[]) {
-    return filters.some(
-      (filter) => filter.conditional === '=' && filter.attribute === value,
-    );
-  }
-
-  function handleNewQuery(value: string) {
-    filter.set({ ...emptyFilter(), attribute: value, conditional: '=' });
-    $focusedElementId = getFocusedElementId(value);
-  }
-
   function updateFocusedElementId() {
     if ($activeQueryIndex !== null) {
       $focusedElementId = getFocusedElementId($filter.attribute);
@@ -142,18 +121,6 @@
   }
 
   $: $activeQueryIndex, updateFocusedElementId();
-
-  let searchAttributeValue = '';
-  //  TODO: Add KeywordList support
-  $: options = $sortedSearchAttributeOptions.filter(
-    (option) => !isListFilter(option.value),
-  );
-
-  $: filteredOptions = !searchAttributeValue
-    ? options
-    : options.filter((option) =>
-        option.value.toLowerCase().includes(searchAttributeValue.toLowerCase()),
-      );
 
   function updateFocus() {
     if ($focusedElementId) {
@@ -174,7 +141,6 @@
   function resetFilter() {
     activeQueryIndex.set(null);
     filter.set(emptyFilter());
-    searchAttributeValue = '';
   }
 
   function handleKeyUp(event: KeyboardEvent) {
@@ -197,47 +163,7 @@
         {#if isStatusFilter($filter.attribute)}
           <StatusFilter />
         {:else}
-          <MenuContainer>
-            <MenuButton
-              controls="search-attribute-menu"
-              unroundRight={Boolean($filter.attribute)}
-              disabled={$activeQueryIndex !== null}
-              count={$filter.attribute ? 0 : $workflowFilters.length}
-            >
-              <svelte:fragment slot="leading">
-                {#if !$filter.attribute}
-                  <Icon name="filter" />
-                {/if}
-              </svelte:fragment>
-              {$filter.attribute || translate('workflows', 'filter')}
-            </MenuButton>
-            <Menu id="search-attribute-menu">
-              <Input
-                label={translate('search')}
-                labelHidden
-                id="filter-search"
-                noBorder
-                bind:value={searchAttributeValue}
-                icon="search"
-                placeholder={translate('search')}
-                class="mb-1"
-              />
-
-              {#each filteredOptions as { value, label }}
-                {@const disabled = isOptionDisabled(value, $workflowFilters)}
-                <MenuItem
-                  on:click={() => {
-                    handleNewQuery(value);
-                  }}
-                  {disabled}
-                >
-                  {label}
-                </MenuItem>
-              {:else}
-                <MenuItem on:click={noop}>{translate('no-results')}</MenuItem>
-              {/each}
-            </Menu>
-          </MenuContainer>
+          <SearchAttributeMenu />
         {/if}
 
         {#if isTextFilter($filter.attribute)}
